+++ condition-order-domain/src/test/java/me/caosh/condition/domain/model/strategy/container/TestPriceStrategy.java
@@ -35,7 +35,7 @@ class TestPriceStrategy implements MarketDrivenStrategy {
     }
 
     @Override
-    public int getStrategyId() {
+    public long getStrategyId() {
         return strategyId;
     }
