+++ condition-order-domain/src/test/java/me/caosh/condition/domain/model/strategy/container/TestTurnUpStrategy.java
@@ -25,7 +25,7 @@ public class TestTurnUpStrategy implements MarketDrivenStrategy, DirtyFlag {
         this.turnUpCondition = turnUpCondition;
     }
 
-    public int getStrategyId() {
+    public long getStrategyId() {
         return strategyId;
     }
